move healthchecks outside of packager
